requirements.txt
flask==2.3.2
tensorflow-cpu==2.12.1  # Last reliably available CPU version
numpy==1.23.5  # TF 2.12's preferred version
scikit-learn==1.2.2
joblib==1.2.0
gunicorn==21.2.0
protobuf==3.20.3
setuptools==65.5.0